refactor(core): the two members that were never transition machinery move out (#1607)
`navigateToNotFound` bypasses the pipeline entirely — it hand-builds an
UNKNOWN_ROUTE state, freezes it, commits it and emits only TRANSITION_SUCCESS.
The one thing it shares with a transition is supersession, so it takes
`abortPreviousNavigation` with it and nothing else. `resolveAsyncGuard` turns a
guard's rejection or `false` into the phase's `RouterError`, which is
`errorHandling.ts`'s whole concern, not the interpreter's.

`InFlightNavigation.ts` is PascalCase per the repo's class-file convention
(`EventEmitter.ts`, `NavigationNamespace.ts`), and stays beside the namespace
rather than under `transition/`: its state outlives any single transition — the
token is monotonic across navigations, and the FSM CANCEL action reaches it from
outside a transition entirely.

NavigationNamespace.ts is 327 lines (from 962). Docs synced: the file map, the
adoption rule for the controller, and every stale `#privateMethod` reference in
CLAUDE.md. 3840/432/153 green, coverage 100%.
