+++ packages/core/src/namespaces/NavigationNamespace/InFlightNavigation.ts
