+++ packages/core/src/namespaces/NavigationNamespace/NavigationNamespace.ts
@@ -5,20 +5,12 @@ import {
   isExpectedRejection,
   PRE_SUPPRESSED,
 } from "./constants";
-import { InFlightNavigation } from "./inFlightNavigation";
-import {
-  abortPreviousNavigation,
-  executeNavigation,
-} from "./transition/executeNavigation";
+import { InFlightNavigation } from "./InFlightNavigation";
+import { executeNavigation } from "./transition/executeNavigation";
+import { navigateToNotFound } from "./transition/navigateToNotFound";
 import { findMisChanneledKey, misChanneledKeyMessage } from "../../channels";
-import {
-  EMPTY_PARAMS,
-  EMPTY_SEARCH,
-  errorCodes,
-  constants,
-} from "../../constants";
+import { errorCodes, constants } from "../../constants";
 import { RouterError } from "../../RouterError";
-import { nameToIDs } from "../../transitionPath";
 
 import type { NavigationDependencies } from "./types";
 import type {
@@ -26,13 +18,7 @@ import type {
   Params,
   SearchParams,
   State,
-  TransitionMeta,
 } from "../../types";
-
-const FROZEN_ACTIVATED: string[] = Object.freeze([
-  constants.UNKNOWN_ROUTE,
-]) as unknown as string[];
-const FROZEN_REPLACE_OPTS: NavigationOptions = Object.freeze({ replace: true });
 
 /**
  * Independent namespace for managing navigation.
@@ -114,63 +100,7 @@ export class NavigationNamespace {
   }
 
   navigateToNotFound(path: string): State {
-    // #1186 — liveness gate. This internal commit primitive has no FSM
-    // transition of its own, so without this check a `dispose()` that lands
-    // while a start-interceptor is parked (FSM already DISPOSED) would let the
-    // resuming pipeline commit an UNKNOWN_ROUTE state on the disposed router and
-    // `start()` resolve. Symmetric with `navigateToState`'s `canNavigate()` gate
-    // (the matched-route branch is already protected). `!isActive()` also covers
-    // a merely-stopped (IDLE) router: the only reachable path to that is a direct
-    // `router.navigateToNotFound()` on a stopped instance (internal callers run
-    // during STARTING, which is active), so the ROUTER_DISPOSED code is slightly
-    // broad there — fail-closed is deliberate (committing on a stopped router is
-    // out of contract), and the disposed race is the case that matters.
-    if (!this.#deps.isActive()) {
-      throw new RouterError(errorCodes.ROUTER_DISPOSED);
-    }
-
-    abortPreviousNavigation(this.#deps);
-
-    const fromState = this.#deps.getState();
-    const deactivated: string[] = fromState
-      ? nameToIDs(fromState.name).toReversed()
-      : [];
-
-    Object.freeze(deactivated);
-
-    const segments: TransitionMeta["segments"] = {
-      deactivated,
-      activated: FROZEN_ACTIVATED,
-      intersection: "",
-    };
-
-    Object.freeze(segments);
-
-    const transitionMeta: TransitionMeta = {
-      phase: "activating",
-      ...(fromState && { from: fromState.name }),
-      reason: "success",
-      replace: true,
-      segments,
-    };
-
-    Object.freeze(transitionMeta);
-
-    const state: State = {
-      name: constants.UNKNOWN_ROUTE,
-      params: EMPTY_PARAMS,
-      search: EMPTY_SEARCH,
-      path,
-      transition: transitionMeta,
-      context: {},
-    };
-
-    Object.freeze(state);
-
-    this.#deps.setState(state);
-    this.#deps.emitTransitionSuccess(state, fromState, FROZEN_REPLACE_OPTS);
-
-    return state;
+    return navigateToNotFound(this.#deps, path);
   }
 
   /**
